fix percentage capacity

+++ db/migrate/20161014184202_create_reservoirs.rb
@@ -2,8 +2,8 @@ class CreateReservoirs < ActiveRecord::Migration
   def change
     create_table :reservoirs do |t|
       t.string   :name
-      t.float    :storage
-      t.float    :capacity
+      t.integer  :storage
+      t.integer  :capacity
 
       t.timestamps(null: false)
     end
